content/canvas/test/test_2d.path.isPointInPath.bezier.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.path.isPointInPath.bezier</title>
<!-- Testing: isPointInPath() works on Bezier curves -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.moveTo(25, 25);
ctx.bezierCurveTo(50, -50, 50, 100, 75, 25);
ok(ctx.isPointInPath(25, 20) == false, "ctx.isPointInPath(25, 20) == false");
ok(ctx.isPointInPath(25, 30) == false, "ctx.isPointInPath(25, 30) == false");
ok(ctx.isPointInPath(30, 20) == true, "ctx.isPointInPath(30, 20) == true");
ok(ctx.isPointInPath(30, 30) == false, "ctx.isPointInPath(30, 30) == false");
ok(ctx.isPointInPath(40, 2) == false, "ctx.isPointInPath(40, 2) == false");
ok(ctx.isPointInPath(40, 20) == true, "ctx.isPointInPath(40, 20) == true");
ok(ctx.isPointInPath(40, 30) == false, "ctx.isPointInPath(40, 30) == false");
ok(ctx.isPointInPath(40, 47) == false, "ctx.isPointInPath(40, 47) == false");
ok(ctx.isPointInPath(45, 20) == true, "ctx.isPointInPath(45, 20) == true");
ok(ctx.isPointInPath(45, 30) == false, "ctx.isPointInPath(45, 30) == false");
ok(ctx.isPointInPath(55, 20) == false, "ctx.isPointInPath(55, 20) == false");
ok(ctx.isPointInPath(55, 30) == true, "ctx.isPointInPath(55, 30) == true");
ok(ctx.isPointInPath(60, 2) == false, "ctx.isPointInPath(60, 2) == false");
ok(ctx.isPointInPath(60, 20) == false, "ctx.isPointInPath(60, 20) == false");
ok(ctx.isPointInPath(60, 30) == true, "ctx.isPointInPath(60, 30) == true");
ok(ctx.isPointInPath(60, 47) == false, "ctx.isPointInPath(60, 47) == false");
ok(ctx.isPointInPath(70, 20) == false, "ctx.isPointInPath(70, 20) == false");
ok(ctx.isPointInPath(70, 30) == true, "ctx.isPointInPath(70, 30) == true");
ok(ctx.isPointInPath(75, 20) == false, "ctx.isPointInPath(75, 20) == false");
ok(ctx.isPointInPath(75, 30) == false, "ctx.isPointInPath(75, 30) == false");

SimpleTest.finish();

});
</script>
